refactor&feat: improved indexing function & improved indexedToNode map creation
index to node map no longer relies on storing data.
used store to keep track of current page index and used useEffect to create index to node map when a pages index equals current page index (stored in store as readingPageIndex). This lays the ground work for a reading controller